[fix] IO#pread/pwrite: use long for file offset to support >2GB positions
pread and pwrite used int (32-bit) for the file offset parameter,
causing RangeError for positions beyond ~2.1GB. CRuby uses off_t
(64-bit) and handles multi-GB file positions correctly.

Changed the offset type from int to long through the call chain:
- RubyIO.pread/pwrite: toInt → toLong for offset argument
- OpenFile.preadInternal/pwriteInternal: int from → long from
- OpenFile PREAD_TASK/PWRITE_TASK: added long run() overload
- RubyThread.ReadWrite: added default long run() method
- RubyThread.executeReadWrite: added long start overload
- FiberScheduler: int from → long from for pread/pwrite methods

Preserves the executeReadWrite threading architecture
(preReadWrite/postReadWrite for interruptibility) by adding long
overloads rather than bypassing the framework.

The underlying Java APIs (FileChannel.read/write and JNR POSIX
pread/pwrite) already accept long positions.

+++ core/src/main/java/org/jruby/FiberScheduler.java
@@ -74,7 +74,7 @@ public class FiberScheduler {
     }
 
     // MRI: rb_fiber_scheduler_io_pread
-    public static IRubyObject ioPRead(ThreadContext context, IRubyObject scheduler, IRubyObject io, IRubyObject buffer, int from, int length, int offset) {
+    public static IRubyObject ioPRead(ThreadContext context, IRubyObject scheduler, IRubyObject io, IRubyObject buffer, long from, int length, int offset) {
         return Helpers.invokeChecked(context, scheduler, "io_pread", io, buffer, asFixnum(context, from), asFixnum(context, length), asFixnum(context, offset));
     }
 
@@ -92,7 +92,7 @@ public class FiberScheduler {
     }
 
     // MRI: rb_fiber_scheduler_io_pwrite
-    public static IRubyObject ioPWrite(ThreadContext context, IRubyObject scheduler, IRubyObject io, IRubyObject buffer, int from, int length, int offset) {
+    public static IRubyObject ioPWrite(ThreadContext context, IRubyObject scheduler, IRubyObject io, IRubyObject buffer, long from, int length, int offset) {
         return Helpers.invokeChecked(context, scheduler, "io_pwrite", io, buffer, asFixnum(context, from), asFixnum(context, length), asFixnum(context, offset));
     }
 
@@ -113,7 +113,7 @@ public class FiberScheduler {
     }
 
     // MRI: rb_fiber_scheduler_io_pread_memory
-    public static IRubyObject ioPReadMemory(ThreadContext context, IRubyObject scheduler, IRubyObject io, ByteBuffer base, int from, int size, int length) {
+    public static IRubyObject ioPReadMemory(ThreadContext context, IRubyObject scheduler, IRubyObject io, ByteBuffer base, long from, int size, int length) {
         RubyIOBuffer buffer = RubyIOBuffer.newBuffer(context, base, size, RubyIOBuffer.LOCKED);
 
         IRubyObject result = ioPRead(context, scheduler, io, buffer, from, length, 0);
@@ -137,7 +137,7 @@ public class FiberScheduler {
     }
 
     // MRI: p
-    public static IRubyObject ioPWriteMemory(ThreadContext context, IRubyObject scheduler, IRubyObject io, ByteBuffer base, int from, int size, int length) {
+    public static IRubyObject ioPWriteMemory(ThreadContext context, IRubyObject scheduler, IRubyObject io, ByteBuffer base, long from, int size, int length) {
         RubyIOBuffer buffer = RubyIOBuffer.newBuffer(context, base, size, RubyIOBuffer.LOCKED | RubyIOBuffer.READONLY);
 
         IRubyObject result = ioPWrite(context, scheduler, io, buffer, from, length, 0);
